ext/gd/tests: fine-grained checks for GD/GD2 support
When using an external libgd, support for the "GD" and "GD2" image
formats is optional. Until now, we were unable to detect their
presence from within the test suite, so the next best thing was done:
a version check for external v2.3.3, which disabled those formats by
default.

But now, PHP has proper format detection and ifdefs around all of the
corresponding functions such as imagegd2(). This commit removes the
old version checks from the test suite, and replaces them with proper
checks for imagegd2() and friends.