ext/gd/tests/bug73869.phpt: ignore warnings from external libgd
If an external libgd is built with support for the GD/GD2 formats,
this test will raise two additional memory warnings in addition to
expected "not a valid GD2 file." This is not a problem, and the
upstream commit responsible for the warnings will probably find its
way into our bundled libgd, so for now we just ignore the additional
warnings. If/when the bundled libgd is updated, they can be made
unconditionally expected.