ext/session/mod_mm.c: add a few missing ZSTR macros
In eaee504c the session's save_path global was changed to a
zend_string pointer, but there are a few direct char-pointer accesses
in ext/session/mod_mm.c that slipped through the cracks. GCC-15
notices them and fails to build due to the incompatible pointer types.
Three ZSTR_* wrappers is all that is needed.

Gentoo-Bug: https://bugs.gentoo.org/967862